Move Tablinks to shared-module
