+++ services/main-frontend/src/pages/manage/courses/[id]/[...path].tsx
@@ -2,9 +2,6 @@ import React from "react"
 import { useTranslation } from "react-i18next"
 
 import Layout from "../../../../components/Layout"
-import TabLink from "../../../../components/TabLink"
-import TabLinkNavigation from "../../../../components/TabLinkNavigation"
-import TabLinkPanel from "../../../../components/TabLinkPanel"
 import CourseChangeRequests from "../../../../components/page-specific/manage/courses/id/change-request/CourseChangeRequests"
 import CourseCourseInstances from "../../../../components/page-specific/manage/courses/id/course-instances/CourseCourseInstances"
 import CourseExercises from "../../../../components/page-specific/manage/courses/id/exercises/CourseExercises"
@@ -17,6 +14,9 @@ import CoursePermissions from "../../../../components/page-specific/manage/cours
 import CourseStatsPage from "../../../../components/page-specific/manage/courses/id/stats/CourseStatsPage"
 import createPendingChangeRequestCountHook from "../../../../hooks/count/usePendingChangeRequestCount"
 import createUnreadFeedbackCountHook from "../../../../hooks/count/useUnreadFeedbackCount"
+import TabLink from "../../../../shared-module/components/Navigation/TabLinks/TabLink"
+import TabLinkNavigation from "../../../../shared-module/components/Navigation/TabLinks/TabLinkNavigation"
+import TabLinkPanel from "../../../../shared-module/components/Navigation/TabLinks/TabLinkPanel"
 import { withSignedIn } from "../../../../shared-module/contexts/LoginStateContext"
 import useQueryParameter from "../../../../shared-module/hooks/useQueryParameter"
 import {
