+++ shared-module/src/locales/en/main-frontend.json
@@ -207,7 +207,6 @@
   "text-field-label-path": "Path",
   "text-field-label-title": "Title",
   "unread": "Unread",
-  "tab-aria-label-default": "Choose tab to display",
   "text-visible-when-feedback-given": "Text visible when the feedback was given",
   "this-field-required": "This field is required",
   "title-all-course-instances": "All course instances",
