+++ shared-module/src/locales/fi/main-frontend.json
@@ -205,7 +205,6 @@
   "text-field-label-path": "Polku",
   "text-field-label-title": "Otsikko",
   "unread": "Lukemattomat",
-  "tab-aria-label-default": "Valitse näytettävä välilehti",
   "text-visible-when-feedback-given": "Teksti näkyvillä kun palaute oli annettu",
   "this-field-required": "Tämä kenttä on vaadittu",
   "title-all-course-instances": "Kaikki kurssin versiot",
