+++ shared-module/src/locales/fi/shared-module.json
@@ -28,6 +28,7 @@
   "error-title": "Virhe",
   "faq": "UKK",
   "hide-trace": "Piilota trace",
+  "tab-aria-label-default": "Valitse näytettävä välilehti",
   "home-page": "Kotisivu",
   "license": "Lisenssi",
   "log-in": "Kirjaudu sisään",
